Move IPv6 bogon to bogon
Not sure if this request will be accepted. Would it not be ideal to have this file for both v4 and v6 bogons? I think there is some use in an IPv4 bogon configuration.

If so, we could either have a separate v4 file, or merge both v4 and v6 into a single bogon file.